# apps/api/openapi.yaml
openapi: 3.0.3
info:
  title: HomesByGurmeet API
  version: 1.0.0
  description: |
    REST API for public property browsing and lead capture, plus the authenticated
    HomesByGurmeet staff dashboard. The API is served by Express on port 4000.
    
    Admin sessions use the HTTP-only `hbg_session` cookie issued by the login endpoint.
servers:
  - url: http://localhost:4000/api
    description: Local Docker development API
...
paths:
...
  /properties:
    get:
      tags: [Properties]
      summary: List public properties
      security: []
      parameters:
        - $ref: '#/components/parameters/StatusQuery'
        - $ref: '#/components/parameters/SideQuery'
        - name: source
          in: query
          schema: { type: string, enum: [redfin, zillow] }
        - name: sort
          in: query
          schema: { type: string, enum: [price_desc, price_asc] }
        - name: page
          in: query
          schema: { type: integer, minimum: 1, default: 1 }
      responses:
        '200':
          description: Paginated public properties
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedProperties'
...
  /properties/{slug}:
    get:
      tags: [Properties]
      summary: Get a public property by URL slug
      security: []
      parameters:
        - $ref: '#/components/parameters/PropertySlug'
      responses:
        '200':
          description: Property details
          content:
            application/json:
              schema: { $ref: '#/components/schemas/PublicProperty' }
        '404': { $ref: '#/components/responses/NotFound' }
  /leads:
    post:
      tags: [Leads]
      summary: Submit a public lead form
      security: []
      description: |
        Creates an inquiry, showing request, valuation request, or general contact lead.
        Requests are rate limited to 10 submissions per 15 minutes per client. If Turnstile is configured,
        `turnstileToken` is required and verified server-side.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CreateLead' }
      responses:
        '201':
          description: Lead stored
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Ok' }
        '422': { $ref: '#/components/responses/ValidationError' }
        '429':
          description: Submission rate limit exceeded
  /analytics/events:
    post:
      tags: [Public]
      summary: Record an anonymous page or listing view
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [type, path]
              properties:
                type: { type: string, enum: [page_view, listing_view] }
                path: { type: string, example: /homes-for-sale }
                propertyId: { type: string, pattern: '^[a-fA-F0-9]{24}$' }
      responses:
        '204': { description: Analytics event recorded }
        '422': { $ref: '#/components/responses/ValidationError' }
  /auth/login:
    post:
      tags: [Authentication]
      summary: Sign in a staff account
      security: []
      description: Sets an HTTP-only `hbg_session` cookie valid for eight hours.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password]
              properties:
                email: { type: string, format: email }
                password: { type: string, format: password }
      responses:
        '200':
          description: Authenticated staff account
          headers:
            Set-Cookie:
              schema: { type: string }
              description: HTTP-only `hbg_session` session cookie
          content:
            application/json:
              schema:
                type: object
                properties:
                  user: { $ref: '#/components/schemas/CurrentUser' }
        '401': { $ref: '#/components/responses/Unauthorized' }
  /auth/logout:
    post:
      tags: [Authentication]
      summary: End the current staff session
      security: [{ CookieAuth: [] }]
      responses:
        '200':
          description: Session cookie cleared
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Ok' }
  /auth/me:
    get:
      tags: [Authentication]
      summary: Get the authenticated staff account
      security: [{ CookieAuth: [] }]
      responses:
        '200':
          description: Current user
          content:
            application/json:
              schema: { $ref: '#/components/schemas/CurrentUser' }
        '401': { $ref: '#/components/responses/Unauthorized' }
  /admin/properties:
    get:
      tags: [Staff administration]
      summary: List all properties for staff management
      security: [{ CookieAuth: [] }]
      responses:
        '200':
          description: All property records, including source and override metadata
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/Property' }
        '401': { $ref: '#/components/responses/Unauthorized' }
    post:
      tags: [Staff administration]
      summary: Create a manual property
      security: [{ CookieAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CreateProperty' }
      responses:
        '201':
          description: Property created
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Property' }
        '401': { $ref: '#/components/responses/Unauthorized' }
  /admin/properties/{id}:
    patch:
      tags: [Staff administration]
      summary: Update a property
      description: Send only the fields to update. Set `manualOverrides` to preserve manually edited price, description, or images during a later data sync.
      security: [{ CookieAuth: [] }]
      parameters: [{ $ref: '#/components/parameters/ObjectId' }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/UpdateProperty' }
      responses:
        '200':
          description: Updated property
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Property' }
        '401': { $ref: '#/components/responses/Unauthorized' }
    delete:
      tags: [Staff administration]
      summary: Delete a property
      security: [{ CookieAuth: [] }]
      parameters: [{ $ref: '#/components/parameters/ObjectId' }]
      responses:
        '204': { description: Property deleted }
        '401': { $ref: '#/components/responses/Unauthorized' }
  /admin/uploads:
    post:
      tags: [Staff administration]
      summary: Upload listing images
      description: Upload up to 12 image files; each file may be at most 8 MB. Use the returned paths in a property `images` array.
      security: [{ CookieAuth: [] }]
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [images]
              properties:
                images:
                  type: array
                  maxItems: 12
                  items: { type: string, format: binary }
      responses:
        '200':
          description: Managed image paths
          content:
            application/json:
              schema:
                type: object
                properties:
                  paths:
                    type: array
                    items: { type: string, example: /uploads/497ec3e1af85d3b7 }
        '401': { $ref: '#/components/responses/Unauthorized' }
  /admin/leads:
    get:
      tags: [Staff administration]
      summary: List submitted leads
      security: [{ CookieAuth: [] }]
      responses:
        '200':
          description: Leads newest first
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/Lead' }
        '401': { $ref: '#/components/responses/Unauthorized' }
  /admin/logs:
    get:
      tags: [Staff administration]
      summary: List recent sync logs
      security: [{ CookieAuth: [] }]
      responses:
        '200':
          description: Up to 30 logs, newest first
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/SystemLog' }
        '401': { $ref: '#/components/responses/Unauthorized' }
  /admin/analytics:
    get:
      tags: [Staff administration]
      summary: Get listing, lead, traffic, and data-health analytics
      security: [{ CookieAuth: [] }]
      parameters:
        - name: from
          in: query
          schema: { type: string, format: date, example: 2026-09-01 }
        - name: to
          in: query
          schema: { type: string, format: date, example: 2026-10-04 }
        - name: range
          in: query
          description: Use all for all retained analytics data.
          schema: { type: string, enum: [all] }
      responses:
        '200':
          description: Aggregated analytics response
          content:
            application/json:
              schema:
                type: object
                properties:
                  range: { type: object }
                  overview: { type: object }
                  traffic: { type: object }
                  leads: { type: object }
                  inventory: { type: object }
                  dataHealth: { type: object }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '422': { $ref: '#/components/responses/ValidationError' }
  /admin/listings/import:
    post:
      tags: [Staff administration]
      summary: Atomically import paired Redfin and Zillow workbooks
      description: |
        Upload both .xlsx workbooks together. Both are validated before one transaction
        replaces their snapshots and rebuilds the public Property collection. Every valid
        source row is retained, including duplicate addresses.
      security: [{ CookieAuth: [] }]
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [redfinFile, zillowFile]
              properties:
                redfinFile:
                  type: string
                  format: binary
                  description: Redfin .xlsx workbook
                zillowFile:
                  type: string
                  format: binary
                  description: Zillow .xlsx workbook
      responses:
        '200':
          description: Workbook imported and listings rebuilt
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok: { type: boolean, example: true }
                  sources: { type: array, items: { type: object } }
                  totalListings: { type: integer }
                  created: { type: integer }
                  updated: { type: integer }
                  removed: { type: integer }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '422': { $ref: '#/components/responses/ValidationError' }
  /admin/users:
    get:
      tags: [Staff administration]
      summary: List staff accounts
      security: [{ CookieAuth: [] }]
      responses:
        '200':
          description: Staff accounts
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/StaffUser' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
    post:
      tags: [Staff administration]
      summary: Create an editor or administrator
      security: [{ CookieAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password]
              properties:
                email: { type: string, format: email }
                password: { type: string, format: password, minLength: 8 }
                role: { $ref: '#/components/schemas/UserRole' }
      responses:
        '201':
          description: User created
          content:
            application/json:
              schema: { $ref: '#/components/schemas/StaffUser' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
...
components:
  securitySchemes:
    CookieAuth:
      type: apiKey
      in: cookie
      name: hbg_session
      description: HTTP-only JWT cookie returned by POST /auth/login.
  parameters:
    PropertySlug:
      name: slug
      in: path
      required: true
      schema: { type: string, example: 123-example-street }
...
    StatusQuery:
      name: status
      in: query
      schema: { type: string, enum: [Active, Pending, Sold] }
    SideQuery:
      name: side
      in: query
      description: Filter sold properties by Gurmeet's transaction representation.
      schema: { type: string, enum: [buyer, seller] }
  responses:
    Unauthorized:
      description: Missing, expired, or invalid session cookie
...
    NotFound:
      description: Resource was not found
...
  schemas:
    Ok:
      type: object
      properties: { ok: { type: boolean, example: true } }
...
    PaginatedProperties:
      type: object
      properties:
        items: { type: array, items: { $ref: '#/components/schemas/PublicProperty' } }
        total: { type: integer }
        page: { type: integer }
        pages: { type: integer }
    CreateLead:
      type: object
      required: [type, name, email, phone]
      properties:
        type: { type: string, enum: [inquiry, showing, valuation, contact] }
        name: { type: string, minLength: 2 }
        email: { type: string, format: email }
        phone: { type: string, minLength: 6 }
        message: { type: string, maxLength: 2000 }
        property: { type: string, description: MongoDB property ID }
        preferredDate: { type: string, format: date-time }
        propertyAddress: { type: string }
        squareFootage: { type: number }
        condition: { type: string }
        turnstileToken: { type: string, description: Required only when Turnstile is configured. }
...